Added getConfig method
To assist with troubleshooting, Cas::getConfig() has been added for determining runtime configuration related issues.  Also, when logging is enabled, the configuration keys and values are now written as they are loaded.  Modified logout to continue logout even when session data not present.  If logging is enabled, it will be written to the log.