Updated demo middleware
Improved code accuracy for proper workflow.  phpCAS uses it's session and ticketing system for doing authentication checks.  It is possible for Laravel sessions to expire before a TGT can expire, thus we move the Laravel session handling to the Guard implementation and not the CAS middleware.  The Guard or Authorization Middleware can leverage from `cas()->user()` to retrieve further credentials for authorization and application login.