Update cas proxy signatures (#136)
* Use phpDoc type hints compatible with apereo/phpcas methods
* Allow CAS_CLIENT_SERVICE to be given as comma separated string

The phpCAS methods allow an array of URLs to be given for the service base. The previous implementation unnecessarily limited this package to a single string value. The package config file will now explode the value on commas, resulting in an array of possible values.

* Move cas_client_service split from config.php into CasManager::ParseConfig

This allows us to pass a simple string into the phpCAS Client and Proxy methods, thereby avoiding autodiscovery. A comma separated list will be detected and split into an array when appropriate.

* Ensure array index exists before trying to use it

---------

Co-authored-by: Dirk Tepe <tepeds@miamioh.edu>